backend/pyproject.toml
[tool.poetry]
name = "ewarehouse"
version = "0.1.0"
description = ""
authors = ["serhatuluc <ulucs15@itu.edu.tr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
bandit = "^1.7.8"
safety = "^3.1.0"
ruff = "^0.4.2"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
